Keep nll_loss_backward's decomposition in self's dtype
The ATen kernel returns self.dtype, but the decomposition let type promotion
escape into the result: with reduction=mean, grad_output / total_weight
promotes to the wider of the two, and the final grad_input * grad_output
inherits it. An fp16 input with an fp32 total_weight came back fp32.

Values were already correct -- only the dtype was wrong -- so cast the result
back rather than changing the arithmetic.

Device-independent (nll_loss2d_backward shares this helper). Verified on MPS:
fp16 now round-trips with 0.0 difference against eager, and the three
nll_loss inductor tests pass.

+++ torch/_decomp/decompositions.py
@@ -533,7 +533,10 @@ def _nll_loss_backward(
 
     grad_output = torch.where(target != ignore_index, grad_output, 0)
 
-    return grad_input * grad_output
+    # The ATen kernel returns self.dtype. A total_weight or weight in a wider
+    # dtype promotes grad_output above it, so cast back rather than letting the
+    # promotion escape into the result.
+    return (grad_input * grad_output).to(self.dtype)
 
 
 @register_decomposition(aten.glu_backward)
